third commit

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,17 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -72,6 +83,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post =  Post::find($id);
+        //check for correct user
+        if(auth()->id !== $post->user_id) {
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -107,6 +122,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+          //check for correct user
+        if(auth()->id !== $post->user_id) {
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
